build & apply transformation matrix
Node tree is traversed to build a transformation matrix for the first node that instantiates the mesh.
This matrix is applied to all vertices.
The transpose of the inverse is applied to normal & tangent.
When the determinant of the matrix is negative renormalisation is required,
the bitangent is inverted and the winding order is revered.